requirements.txt
pytest
pytest-cov
pydantic
pyyaml
mcp
fastapi
uvicorn
httpx
